Improve clean up if extraction fails
