+++ lib/setup-codeql.js
@@ -38,7 +38,6 @@ const fs = __importStar(require("fs"));
 const path = __importStar(require("path"));
 const perf_hooks_1 = require("perf_hooks");
 const toolcache = __importStar(require("@actions/tool-cache"));
-const del_1 = __importDefault(require("del"));
 const fast_deep_equal_1 = __importDefault(require("fast-deep-equal"));
 const semver = __importStar(require("semver"));
 const uuid_1 = require("uuid");
@@ -403,12 +402,18 @@ const downloadCodeQL = async function (codeqlURL, maybeBundleVersion, maybeCliVe
     const archivedBundlePath = await toolcache.downloadTool(codeqlURL, dest, authorization, finalHeaders);
     const downloadDurationMs = Math.round(perf_hooks_1.performance.now() - toolsDownloadStart);
     logger.debug(`Finished downloading CodeQL bundle to ${archivedBundlePath} (${downloadDurationMs} ms).`);
-    logger.debug("Extracting CodeQL bundle.");
-    const extractionStart = perf_hooks_1.performance.now();
-    const extractedBundlePath = await tar.extract(archivedBundlePath, compressionMethod, tarVersion);
-    const extractionDurationMs = Math.round(perf_hooks_1.performance.now() - extractionStart);
-    logger.debug(`Finished extracting CodeQL bundle to ${extractedBundlePath} (${extractionDurationMs} ms).`);
-    await cleanUpGlob(archivedBundlePath, "CodeQL bundle archive", logger);
+    let extractedBundlePath;
+    let extractionDurationMs;
+    try {
+        logger.debug("Extracting CodeQL bundle.");
+        const extractionStart = perf_hooks_1.performance.now();
+        extractedBundlePath = await tar.extract(archivedBundlePath, compressionMethod, tarVersion, logger);
+        extractionDurationMs = Math.round(perf_hooks_1.performance.now() - extractionStart);
+        logger.debug(`Finished extracting CodeQL bundle to ${extractedBundlePath} (${extractionDurationMs} ms).`);
+    }
+    finally {
+        await (0, util_1.cleanUpGlob)(archivedBundlePath, "CodeQL bundle archive", logger);
+    }
     const bundleVersion = maybeBundleVersion ?? tryGetBundleVersionFromUrl(codeqlURL, logger);
     if (bundleVersion === undefined) {
         logger.debug("Could not cache CodeQL tools because we could not determine the bundle version from the " +
@@ -429,7 +434,7 @@ const downloadCodeQL = async function (codeqlURL, maybeBundleVersion, maybeCliVe
     const toolcachedBundlePath = await toolcache.cacheDir(extractedBundlePath, "CodeQL", toolcacheVersion);
     // Defensive check: we expect `cacheDir` to copy the bundle to a new location.
     if (toolcachedBundlePath !== extractedBundlePath) {
-        await cleanUpGlob(extractedBundlePath, "CodeQL bundle from temporary directory", logger);
+        await (0, util_1.cleanUpGlob)(extractedBundlePath, "CodeQL bundle from temporary directory", logger);
     }
     return {
         codeqlFolder: toolcachedBundlePath,
@@ -514,7 +519,7 @@ async function setupCodeQLBundleWithCompressionMethod(toolsInput, apiDetails, te
     switch (source.sourceType) {
         case "local": {
             const compressionMethod = tar.inferCompressionMethod(source.codeqlTarPath);
-            codeqlFolder = await tar.extract(source.codeqlTarPath, compressionMethod, zstdAvailability.version);
+            codeqlFolder = await tar.extract(source.codeqlTarPath, compressionMethod, zstdAvailability.version, logger);
             toolsSource = ToolsSource.Local;
             break;
         }
@@ -541,24 +546,6 @@ async function setupCodeQLBundleWithCompressionMethod(toolsInput, apiDetails, te
         toolsVersion,
         zstdAvailability,
     };
-}
-async function cleanUpGlob(glob, name, logger) {
-    logger.debug(`Cleaning up ${name}.`);
-    try {
-        const deletedPaths = await (0, del_1.default)(glob, { force: true });
-        if (deletedPaths.length === 0) {
-            logger.warning(`Failed to clean up ${name}: no files found matching ${glob}.`);
-        }
-        else if (deletedPaths.length === 1) {
-            logger.debug(`Cleaned up ${name}.`);
-        }
-        else {
-            logger.debug(`Cleaned up ${name} (${deletedPaths.length} files).`);
-        }
-    }
-    catch (e) {
-        logger.warning(`Failed to clean up ${name}: ${e}.`);
-    }
 }
 function sanitizeUrlForStatusReport(url) {
     return ["github/codeql-action", "dsp-testing/codeql-cli-nightlies"].some((repo) => url.startsWith(`https://github.com/${repo}/releases/download/`))
